+++ 2020_1014_02/Practice02/Practice02_04/Enemy.h
@@ -20,9 +20,9 @@ public:
 	void SetMoveSpeed(float speed);
 
 	int GetHp();
-	int GetPosX();
-	int GetPosY();
-	int GetMoveSpeed();
+	float GetPosX();
+	float GetPosY();
+	float GetMoveSpeed();
 
 private:
 	int m_Hp;
