+++ 2020_1014_02/Practice02/Practice02_04/Player.h
@@ -20,15 +20,15 @@ public:
 	void SetMoveSpeed(float speed);
 
 	int GetHp();
-	int GetPosX();
-	int GetPosY();
-	int GetMoveSpeed();
+	float GetPosX();
+	float GetPosY();
+	float GetMoveSpeed();
 
 private:
 	int m_Hp;
-	int m_PosX;
-	int m_PosY;
-	int m_MoveSpeed;
+	float m_PosX;
+	float m_PosY;
+	float m_MoveSpeed;
 };
